[Main] Refactor

+++ online-dashboard-app/src/Components/ProtectedRoute.jsx
@@ -4,15 +4,17 @@ import { AuthContext } from "./AuthContext";
 
 const ProtectedRoute = ({ children, allowedRole }) => {
   const { auth } = useContext(AuthContext);
-  const storedAuth = JSON.parse(localStorage.getItem("data"));
-  const userRole = localStorage.getItem("roles");
   const location = useLocation();
 
-  if (!storedAuth.access_token || !userRole) {
-    return <Navigate to="/" replace />;
+  const storedAuth = JSON.parse(localStorage.getItem("data") || "{}");
+  const userRole = localStorage.getItem("roles");
+
+  if (!storedAuth?.access_token || !userRole) {
+    localStorage.clear();
+    return <Navigate to="/" replace state={{ from: location }} />;
   }
 
-  if (userRole !== allowedRole) {
+  if (allowedRole && userRole !== allowedRole) {
     alert("Unauthorized access. Logging out.");
     localStorage.clear();
     return <Navigate to="/" replace state={{ from: location }} />;
